+++ src/components/status.jsx
@@ -1,10 +1,10 @@
 import './status.css';
 
 import { msg, plural } from '@lingui/core/macro';
-import { Trans, useLingui } from '@lingui/react/macro';
+import { Plural, Trans, useLingui } from '@lingui/react/macro';
 import { ControlledMenu, MenuDivider, MenuItem } from '@szhsin/react-menu';
 import { shallowEqual } from 'fast-equals';
-import pThrottle from 'p-throttle';
+import PQueue from 'p-queue';
 import { Fragment } from 'preact';
 import { memo } from 'preact/compat';
 import {
@@ -30,6 +30,7 @@ import getTranslateTargetLanguage from '../utils/get-translate-target-language';
 import getHTMLText from '../utils/getHTMLText';
 import htmlContentLength from '../utils/html-content-length';
 import localeMatch from '../utils/locale-match';
+import mem from '../utils/mem';
 import niceDateTime from '../utils/nice-date-time';
 import openCompose from '../utils/open-compose';
 import pmem from '../utils/pmem';
@@ -76,20 +77,24 @@ import RelativeTime from './relative-time';
 import StatusButton from './status-button';
 import StatusCard from './status-card';
 import StatusCompact from './status-compact';
+import SubMenu2 from './submenu2';
 import ThreadBadge from './thread-badge';
 import TranslationBlock from './translation-block';
 
 const SHOW_COMMENT_COUNT_LIMIT = 280;
 const INLINE_TRANSLATE_LIMIT = 140;
 
-const throttle = pThrottle({
-  limit: 1,
+const accountQueue = new PQueue({
+  concurrency: 1,
   interval: 1000,
+  intervalCap: 1,
 });
-function fetchAccount(id, masto) {
-  return masto.v1.accounts.$select(id).fetch();
+function fetchAccount(id, masto, signal) {
+  return accountQueue.add(() => masto.v1.accounts.$select(id).fetch(), {
+    signal,
+  });
 }
-const memFetchAccount = pmem(throttle(fetchAccount));
+const memFetchAccount = pmem(fetchAccount);
 
 const isIOS =
   window.ontouchstart !== undefined &&
@@ -164,9 +169,8 @@ function forgivingQSA(selectors = [], dom = document) {
   return [];
 }
 
-function isTranslateble(content, emojis) {
-  if (!content) return false;
-  // Remove custom emojis
+const getHTMLTextForDetectLang = mem((content, emojis) => {
+  if (!content) return '';
   if (emojis?.length) {
     const emojisRegex = new RegExp(
       `:(${emojis.map((e) => e.shortcode).join('|')}):`,
@@ -175,30 +179,7 @@ function isTranslateble(content, emojis) {
     content = content.replace(emojisRegex, '');
   }
   content = content.trim();
-  if (!content) return false;
-  const text = getHTMLText(content, {
-    preProcess: (dom) => {
-      // Remove .mention, pre, code, a:has(.invisible)
-      for (const a of forgivingQSA(
-        ['.mention, pre, code, a:has(.invisible)', '.mention, pre, code'],
-        dom,
-      )) {
-        a.remove();
-      }
-    },
-  });
-  return !!text;
-}
-
-function getHTMLTextForDetectLang(content, emojis) {
-  if (emojis?.length) {
-    const emojisRegex = new RegExp(
-      `:(${emojis.map((e) => e.shortcode).join('|')}):`,
-      'g',
-    );
-    content = content.replace(emojisRegex, '');
-  }
-
+  if (!content) return '';
   return getHTMLText(content, {
     preProcess: (dom) => {
       // Remove anything that can skew the language detection
@@ -223,6 +204,10 @@ function getHTMLTextForDetectLang(content, emojis) {
       }
     },
   });
+});
+
+function isTranslateble(content, emojis) {
+  return !!getHTMLTextForDetectLang(content, emojis);
 }
 
 const SIZE_CLASS = {
@@ -313,6 +298,8 @@ const quoteApprovalPolicyMessages = {
   followers: msg`Your followers can quote`,
   nobody: msg`Only you can quote`,
 };
+
+const QUESTION_REGEX = /[??？︖❓❔⁇⁈⁉¿‽؟]/;
 
 const { DEV } = import.meta.env;
 
@@ -519,19 +506,27 @@ function Status({
     inReplyToAccountRef = { url: accountURL, username, displayName };
   }
   const [inReplyToAccount, setInReplyToAccount] = useState(inReplyToAccountRef);
-  if (!withinContext && !inReplyToAccount && inReplyToAccountId) {
-    const account = states.accounts[inReplyToAccountId];
-    if (account) {
-      setInReplyToAccount(account);
-    } else {
-      memFetchAccount(inReplyToAccountId, masto)
+  useEffect(() => {
+    if (!withinContext && !inReplyToAccount && inReplyToAccountId) {
+      const account = states.accounts[inReplyToAccountId];
+      if (account) {
+        setInReplyToAccount(account);
+        return;
+      }
+
+      const abortController = new AbortController();
+      memFetchAccount(inReplyToAccountId, masto, abortController.signal)
         .then((account) => {
           setInReplyToAccount(account);
           states.accounts[account.id] = account;
         })
         .catch((e) => {});
+
+      return () => {
+        abortController.abort();
+      };
     }
-  }
+  }, [withinContext, inReplyToAccount, inReplyToAccountId]);
   const mentionSelf =
     (inReplyToAccountId && inReplyToAccountId === currentAccount) ||
     mentions?.find((mention) => mention.id === currentAccount);
@@ -754,7 +749,7 @@ function Status({
 
   const postQuoteApprovalPolicy = getPostQuoteApprovalPolicy(quoteApproval);
 
-  const replyStatus = (e) => {
+  const replyStatus = (e, replyMode = 'all') => {
     if (!sameInstance || !authenticated) {
       return alert(unauthInteractionErrorMessage);
     }
@@ -762,11 +757,13 @@ function Status({
     if (e?.shiftKey || e?.syntheticEvent?.shiftKey) {
       const newWin = openCompose({
         replyToStatus: status,
+        replyMode,
       });
       if (newWin) return;
     }
     showCompose({
       replyToStatus: status,
+      replyMode,
     });
   };
 
@@ -1042,17 +1039,85 @@ function Status({
         </div>
       )
     );
+  const mentionsCount = useMemo(() => {
+    if (!mentions?.length) return false;
+    const allMentions = new Set([accountId, ...mentions.map((m) => m.id)]);
+    return [...allMentions].filter((m) => m !== currentAccount).length;
+  }, [accountId, mentions?.length, currentAccount]);
+  const tooManyMentions = mentionsCount > 3;
+  const ReplyMenuContent = () => (
+    <>
+      <Icon icon="comment" />
+      <span>
+        {repliesCount > 0
+          ? shortenNumber(repliesCount)
+          : tooManyMentions
+            ? t`Reply…`
+            : t`Reply`}
+      </span>
+    </>
+  );
+  const replyModeMenuItems = (
+    <>
+      <MenuItem onClick={(e) => replyStatus(e, 'all')}>
+        <small>
+          <Trans>Reply all</Trans>
+          <br />
+          <span class="more-insignificant">
+            <Plural value={mentionsCount} other="# mentions" />
+          </span>
+        </small>
+      </MenuItem>
+      <MenuItem onClick={(e) => replyStatus(e, 'author-first')}>
+        <small>
+          <Trans>Reply all</Trans>
+          <br />
+          <span class="more-insignificant">
+            <Plural
+              value={mentionsCount - 1}
+              other={
+                <Trans comment="Author mention appears first, other mentions appear below with newlines in between">
+                  <span class="bidi-isolate">@{username || acct}</span> first, #
+                  others below
+                </Trans>
+              }
+            />
+          </span>
+        </small>
+      </MenuItem>
+      <MenuItem onClick={(e) => replyStatus(e, 'author-only')}>
+        <small>
+          <Trans>Reply</Trans>
+          <br />
+          <span class="more-insignificant">
+            <Trans>
+              Only <span class="bidi-isolate">@{username || acct}</span>
+            </Trans>
+          </span>
+        </small>
+      </MenuItem>
+    </>
+  );
   const StatusMenuItems = (
     <>
       {!isSizeLarge && sameInstance && (
         <>
           <div class="menu-control-group-horizontal status-menu">
-            <MenuItem onClick={replyStatus}>
-              <Icon icon="comment" />
-              <span>
-                {repliesCount > 0 ? shortenNumber(repliesCount) : t`Reply`}
-              </span>
-            </MenuItem>
+            {tooManyMentions ? (
+              <SubMenu2
+                openTrigger="clickOnly"
+                direction="bottom"
+                overflow="auto"
+                gap={-8}
+                shift={8}
+                menuClassName="menu-emphasized"
+                label={<ReplyMenuContent />}
+              >
+                {replyModeMenuItems}
+              </SubMenu2>
+            ) : (
+              <MenuItem onClick={replyStatus}>{<ReplyMenuContent />}</MenuItem>
+            )}
             <MenuConfirm
               subMenu
               confirmLabel={
@@ -1646,11 +1711,19 @@ function Status({
   );
 
   const hotkeysEnabled = !readOnly && !previewMode && !quoted;
-  const rRef = useHotkeys('r, shift+r', replyStatus, {
-    enabled: hotkeysEnabled,
-    useKey: true,
-    ignoreEventWhen: (e) => e.metaKey || e.ctrlKey || e.altKey,
-  });
+  const rRef = useHotkeys(
+    'r, shift+r',
+    (e, handler) => {
+      // Fix bug: shift+r is fired even when r is pressed due to useKey: true
+      if (e.shiftKey !== handler.shift) return;
+      replyStatus(e);
+    },
+    {
+      enabled: hotkeysEnabled,
+      useKey: true,
+      ignoreEventWhen: (e) => e.metaKey || e.ctrlKey || e.altKey,
+    },
+  );
   const fRef = useHotkeys('f, l', favouriteStatusNotify, {
     enabled: hotkeysEnabled,
     ignoreEventWhen: (e) => e.metaKey || e.ctrlKey || e.altKey || e.shiftKey,
@@ -1856,8 +1929,7 @@ function Status({
     ) {
       return false;
     }
-    const questionRegex = /[??？︖❓❔⁇⁈⁉¿‽؟]/;
-    const containsQuestion = questionRegex.test(content);
+    const containsQuestion = QUESTION_REGEX.test(content);
     if (!containsQuestion) return false;
     if (contentLength > 0 && contentLength <= SHOW_COMMENT_COUNT_LIMIT) {
       return true;
@@ -1995,7 +2067,11 @@ function Status({
                 class="reply-button"
                 icon="comment"
                 iconSize="m"
-                onClick={replyStatus}
+                // Menu doesn't work here
+                // Temporary solution: reply author-first if too many mentions
+                onClick={(e) =>
+                  replyStatus(e, tooManyMentions ? 'author-first' : 'all')
+                }
               />
               <StatusButton
                 size="s"
@@ -2698,14 +2774,36 @@ function Status({
               )}
               <div class={`actions ${_deleted ? 'disabled' : ''}`}>
                 <div class="action has-count">
-                  <StatusButton
-                    title={t`Reply`}
-                    alt={t`Comments`}
-                    class="reply-button"
-                    icon="comment"
-                    count={repliesCount}
-                    onClick={replyStatus}
-                  />
+                  {tooManyMentions ? (
+                    <Menu2
+                      openTrigger="clickOnly"
+                      direction="bottom"
+                      overflow="auto"
+                      gap={-8}
+                      shift={8}
+                      menuClassName="menu-emphasized"
+                      menuButton={
+                        <StatusButton
+                          title={t`Reply`}
+                          alt={t`Comments`}
+                          class="reply-button"
+                          icon="comment"
+                          count={repliesCount}
+                        />
+                      }
+                    >
+                      {replyModeMenuItems}
+                    </Menu2>
+                  ) : (
+                    <StatusButton
+                      title={t`Reply`}
+                      alt={t`Comments`}
+                      class="reply-button"
+                      icon="comment"
+                      count={repliesCount}
+                      onClick={replyStatus}
+                    />
+                  )}
                 </div>
                 {/* <div class="action has-count">
                 <StatusButton